requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic
argon2-cffi
httpx
python-multipart
pypdf
requests
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
